Add files via upload
In this case study, I use Logistic Regression algorithm, which can be used to build models that are simple and yet effective--in some cases. This algorithm can be used to build useful baseline models for classification problems that can be easily interpreted. For these reasons, Logistic Regression is considered as one of the fundamental Machine Learning algorithms associated with classification problems.