Add files via upload
n this case study, I used Random Forest and logistic regression to understand the scope of the Coronavirus using data from December and January of 2020. This case study is an excellent example of how data scientists can help share crucial insights about occurrences that have an impact around the world.